feat(rust-consumer): Move code from rust arroyo to a separate module
We don't want the arroyo code to be in the root of the Snuba project,
move it to a separate module
