+++ rust_snuba/src/arroyo/backends/kafka/errors.rs
@@ -1,6 +1,6 @@
 use rdkafka::error::{KafkaError, RDKafkaErrorCode};
 
-use crate::backends::ConsumerError;
+use crate::arroyo::backends::ConsumerError;
 
 impl From<KafkaError> for ConsumerError {
     fn from(err: KafkaError) -> Self {
