+++ rust_snuba/src/arroyo/examples/base_consumer.rs
