+++ rust_snuba/src/arroyo/examples/base_processor.rs
