+++ rust_snuba/src/arroyo/processing/strategies/mod.rs
@@ -1,4 +1,4 @@
-use crate::types::{Message, Partition, Position};
+use crate::arroyo::types::{Message, Partition, Position};
 use std::collections::HashMap;
 use std::time::Duration;
