--- rust_snuba/src/lib.rs
@@ -1,12 +1,0 @@
-pub mod backends;
-pub mod processing;
-pub mod types;
-pub mod utils;
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
