+++ rust_snuba/src/main.rs
@@ -1,3 +1,5 @@
+mod arroyo;
+
 fn main() {
     println!("Hello, world!");
 }
